added open habits file from the command line.
the expectation is that this feature won't be used often. And, it
creates an issue in that the new journal file currently strips out the
contents of the habits file when making a new entry, meaning changes to
the habits file will mean manual changes to an entry file if the habit
is no longer required.